build template
