+++ bin/template/lib/app.html
@@ -1,5 +1,2 @@
 <h1>My First Angular 2 App</h1>
-<nav>
-    <a [routerLink]="['Home']">Home</a>
-</nav>
 <router-outlet></router-outlet>
