--- bin/template/lib/components/components.dart
@@ -1,1 +1,0 @@
-library __projectName__.components;
