--- bin/template/lib/convert/convert.dart
@@ -1,14 +1,0 @@
-// Generated by Dogma Dart. DO NOT MODIFY!
-// 2016-05-06 21:32:36.052430
-
-library __projectName__.convert.convert;
-
-//---------------------------------------------------------------------
-// Exports
-//---------------------------------------------------------------------
-
-export 'data_convert.dart';
-
-//---------------------------------------------------------------------
-// Library contents
-//---------------------------------------------------------------------
