--- bin/template/lib/convert/data_convert.dart
@@ -1,46 +1,0 @@
-// Generated by Dogma Dart. DO NOT MODIFY!
-// 2016-05-06 21:32:36.052430
-
-library __projectName__.convert.data_convert;
-
-//---------------------------------------------------------------------
-// Standard libraries
-//---------------------------------------------------------------------
-
-import 'dart:convert';
-
-//---------------------------------------------------------------------
-// Imports
-//---------------------------------------------------------------------
-
-import 'package:__projectName__/models.dart';
-import 'package:dogma_convert/convert.dart';
-
-//---------------------------------------------------------------------
-// Library contents
-//---------------------------------------------------------------------
-
-/// A [ModelDecoder] for [Data].
-class DataDecoder extends Converter<Map, Data> implements ModelDecoder<Data> {
-  @override
-  Data create() => new Data();
-  @override
-  Data convert(Map input, [Data model]) {
-    model ??= create();
-
-    model.foo = input['foo'];
-    return model;
-  }
-}
-
-/// A [ModelEncoder] for [Data].
-class DataEncoder extends Converter<Data, Map> implements ModelEncoder<Data> {
-  @override
-  Map convert(Data input) {
-    var model = {};
-
-    model['foo'] = input.foo;
-
-    return model;
-  }
-}
