--- bin/template/lib/directives/directives.dart
@@ -1,1 +1,0 @@
-library __projectName__.directives;
