--- bin/template/lib/models.dart
@@ -1,3 +1,0 @@
-library __projectName__.models;
-
-export 'models/data.dart';
