--- bin/template/lib/models/data.dart
@@ -1,9 +1,0 @@
-/**
- * Created by lejard_h on 06/05/16.
- */
-
-class Data {
-
-    String foo = "bar";
-
-}
