+++ bin/template/lib/pipes.dart
@@ -1,0 +1,1 @@
+
