--- bin/template/lib/pipes/pipes.dart
@@ -1,1 +1,0 @@
-library __projectName__.pipes;
