--- bin/template/lib/routes/routes.dart
@@ -1,3 +1,0 @@
-library __projectName__.routes;
-
-export "home_route/home_route.dart";
