--- bin/template/lib/services/services.dart
@@ -1,1 +1,0 @@
-library __projectName__.services;
