+++ bin/template/tool/grind.dart
@@ -1,8 +1,6 @@
 import "dart:io";
 import "package:grinder/grinder.dart";
 import "package:yaml/yaml.dart";
-import "package:dogma_codegen/build.dart" as dogma_build;
-import "package:watcher/watcher.dart";
 
 class ConfigFile {
   static Map<String, dynamic> _config;
@@ -14,13 +12,6 @@ class ConfigFile {
         _config = loadYaml(_configFile.readAsStringSync()) as Map<String, dynamic>;
       }
     }
-  }
-
-  String get projectName {
-    if (_config != null  && _config.containsKey("project")) {
-      return _config["project"]["name"];
-    }
-    return null;
   }
 
   int get serverPort {
@@ -84,27 +75,7 @@ ConfigFile config = new ConfigFile();
 
 main(List<String> args) async => await grind(args);
 
-@Task("models")
-codegen() async {
-  await dogma_build.build([],
-      modelLibrary: "lib/models.dart",
-      modelPath: "lib/models/",
-      convertPath: "lib/convert",
-      convertLibrary: "lib/convert/convert.dart",
-      mapper: false,
-      unmodifiable: false);
-}
-
-@Task("watchModels")
-watchModels() {
-  new Watcher("lib/models").events.listen((WatchEvent _) {
-    print("[codegen]");
-    codegen();
-  });
-}
-
 @Task("serve")
-@Depends(codegen, watchModels)
 serve() async {
   Process _server = await Process.start(
       "pub",
@@ -157,7 +128,7 @@ test() {
 }
 
 @Task("deploy")
-@Depends(test, doc, codegen, build)
+@Depends(test, doc, build)
 deploy() async {}
 
 @Task("clean")
